Move DetailView examples to site
