+++ packages/terra-site/src/examples/detail-view/DetailViewNoDivider.jsx
@@ -1,14 +1,14 @@
 /* eslint-disable import/no-extraneous-dependencies */
 import React from 'react';
 import Grid from 'terra-grid';
-import DetailView from '../src/DetailView';
+import DetailView from 'terra-detail-view';
 
 const DetailViewDivided = () => (
   <DetailView
     title="Header"
     subtitles={['Subtitle 1', 'Subtitle 2']}
     graph={<div style={{ border: '1px solid black', height: '20em', width: '25em', marginBottom: '0.714rem' }}>This is where any visualizations would go</div>}
-    details={[
+    details={[(
       <Grid key="detailInfo">
         <Grid.Row>
           <Grid.Column col={6}>
@@ -26,9 +26,10 @@ const DetailViewDivided = () => (
             <div>Second Row - Second Column content</div>
           </Grid.Column>
         </Grid.Row>
-      </Grid>,
-    ]}
+      </Grid>
+    )]}
     footer="Footer Text"
+    isDivided={false}
   />
 );
